chg: Renamed kstring Package
Renamed kstring to the "str" package and removed the Str prefix from the
function name.

BREAKING CHANGE
